Refactor sem.c and sem-bare.c to reduce duplication.
This commit moves the implementation of trywait for semaphores into sem-bare.c
to make it available for bare semaphores, and refactors other semaphore
operations into just invocations of their bare semaphore cousins. The purpose
of this is to reduce code duplication for better maintainability. This is not
intended to affect the functionality of semaphores.