Fix: double-word-align stacks.
A thread's stack is expected to be double-word-aligned at ABI boundaries by
various standards (e.g. 5.2.1.2 of the AAPCS [0]). Prior to this commit, we
were only constraining the initial stack to be word-aligned. No fault was
observable in a standard build because GCC page-aligns these regions, though it
is under no obligation to do so.

If your stack is only word-aligned and not double-word-aligned, you will likely
notice nothing unusual unless you are dealing with doubles or floats. In these
circumstances, your compiler may take advantage of stack alignment assumptions
to access these variables in a way that causes a data access fault or a
corrupted read or write.

[0]: http://infocenter.arm.com/help/index.jsp?topic=/com.arm.doc.ihi0042e/index.html