Implement code to examine the stack frame and to print a stack trace
This uses gcc builtins that may or may not be implemented on a given
architecture and which may or may not explode when used. This is
designed as a quick and dirty solution to getting a stack trace if
your code is reaching a place you feel it shouldn't (such as an
error path).