Fix: Make stack trace functionality IA32-only.
The GCC intrinsic __builtin_return_address can only be called with the
parameter 0 on platforms such as ARM that lack a frame pointer. Calling this
with any other value returns garbage. This commit makes the entire
implementation only available on IA32 as even __builtin_return_address(0) is
not helpful on ARM when it will just point at your (known) call site. If we
want this on ARM, we should move this function into a header macro.