Fix incorrect acquisition of a semaphore, due to incorrect waiting.
851502fa64c27d51c23cbdef6b6f09d156b27b12 swapped the atomic decrement that
occurred during waiting on a counting semaphore for a loop that took into
account possible integer overflow. 02e51044c76f0d852fffec9c82eee5d32f898b82 did
the same for binary semaphores. Both of these introduced a bug whereby waiting
on an empty semaphore would incorrectly result in the caller not blocking and
believing they had acquired the lock. This bug resulted in a deadlock in CAmkES
systems.

This commit fixes the bug and makes clearer the semantics of
`sync_atomic_decrement_safe`, namely that the output parameter is the
_previous_ value, not the current value.

Closes JIRA CAMKES-382