Fix memory barriers in semaphores, mutexes and atomic operations
Mutexes and semaphores need to implement acquire and release barriers.
Previously __sync_synchronize was used in a subset of implementations
to provide memory consistency. This change uses specific acquire and
release barriers (which may be more efficient) in all semaphore
and mutex implementations

Additionally the interface of the sync_* functions was changed to take
memory order parameters for passing to the gcc __atomic_* functions.
This allows for much more efficient code as the barriers can just
be done when needed and potentially rolled into the existing atomic
op, providing further efficiencies