Remove 'args' parameter to test cases.
This commit removes the state parameter to test cases, which was unused in all
tests. In three years, no test has required this parameter. I think it's time
to ditch this dead weight. It can be reintroduced if necessary later.

Note that this change is interdependent with changes to sel4test.