libsel4utils: Import page.h definitions from libutils
The page sizes supported by a system are a combination of both the
raw architecture, as well as how seL4 is configured. This makes these
definitions system specific, hence the need to move them

As a consequence of having these defined in libsel4utils it is no
longer possible for libsel4vspace to use them, as libsel4utils
already depends upon libsel4vspace, and we do not support dual
dependencies in the build system. To compensate these checks are
added to the vspace implementation in libsel4utils