libsel4debug: De-frob return addresses before storing.
In the majority of our debugging scenarios, return addresses are not frobbed so
this has not caused problems in the past. However, technically we should always
be extracting the return address before presenting it to the user to avoid
confusion. This commit is primarily for portability and has no expected impact
on typical usage.