Rationalise dependencies.
This commit removes the dependency of libsel4platsupport on
libsel4utils, as libsel4utils is our top level library and should be
able to use the device drivers that libsel4platsupport provides.

The functionality that libsel4platsupport depended on from libsel4utils
has been moved to libsel4simple, as libsel4simple is basically an
extension of libsel4.