sel4utils: don't trample existing cspace
sel4utils_process_configure_custom provides a config
option to allow users provide an existing own cspace.
This change fixes a bug where sel4utils would override
slots in the existing cspace.

The behaviour now is to only write to the processes cspace
if create_cspace is specified in the process conifg.