libsel4utils: Correct check for contiguous page sizes
The intention of the original code was clearly to not check the first
(hence the i > 0 comparison) but then check every subsequent size is
strictly increasing. Instead we skipped the first check and then
compared against the following size, resulting in us never actually
checking if the first page is in order, as well as indexing out
of the array.