sel4utils: Fixed memory leak in vspace impl
As an optimisation, the implementation of vspace in libsel4utils keeps
track of the last vaddr it allocated or freed. This is used as a
starting point for finding regions of unallocated addresses. Prior to
this change, when a range of addresses was freed, this vaddr was set to
the end of the range. This was a problem because newly freed addresses
earlier in the range will never be reused. This fixes the problem by
setting the last allocated/freed vaddr to the start of a newly freed range.