sel4utils: IOMMU DMA alloc for pre-alloc'd buffer
This adds a function to the iommu dma manager that maps frames from an
externally allocated buffer into the iospace. The existing dma
allocation function allocates the buffer from the heap. This makes it
unsuitable in situations without a vspace-backed heap, as caps to the
newly-allocated frames won't be known. Caps to the dma buffer's frames
are needed for mapping the frames into the iospace.