Remove assumptions about the 0 paddr being invalid
There is no way to prevent platforms from having interesting memory
or devices at paddr 0, as such we should not be using 0 to signal
the absence of a physical address. This change fixes some cases
where we were doing this, and switches to sentinal value that cannot
collide with a valid page base physical address.