libsel4allocman: Return allocated objects to the correct pool
The split untyped allocator maintains three pools for different kinds of untyped
objects. Previously the origin of an object was not tracked and when the object
was freed (i.e. returned to the allocator) it was just put back into kernel only
untyped pool. This naturally corrupted the book keeping resulting in future
allocations failing. This change records which pool any allocation is sourced
from and returns back to that pool