sel4utils: remove redundant indirection in timers
This commit removes timer_common_data_t and merges it
with seL4_timer_t, so that irq and vaddr are not hidden
behind an unneccessary layer of indirection. Timers without
a vaddr or irq can just set these to invalid values, however
this change allows users of this library to generically access
these fields.