sel4vmm: Unmapping pages from iospaces
Each guest vspace can have an arbitrary number of iospaces - additional
address spaces used by devices doing dma. When pages are mapped into a
guest vspace, corresponding mappings are added to all of the guest's
iospaces. This change allows these mappings to be removed when unmapping
from a guest vspace.

To avoid leaking cslots, the copies of frame caps used to establish
iospace mappings must now be tracked, so they can be freed when the
mappings are removed. To track the correspondence between guest-physical
addresses and frame caps, an additional vspace is used per iospace.