SELFOUR-572: muslcsys: Init morecore before mmap
It is possible, by doing a large `malloc` that mmap could be called
before muslc ever needs to initialize the morecore region. As our
implementation 'steals' memory from the morecore region for mmap
it will fail if the morecore region is not initialized.

This change factors out the initialization of the morecore region
for both brk and mmap