libsel4platsupport: Refactor HPET interface to support IOAPIC delivery
Previously the HPET interface used the `irq_number` as the delivery
vector and would look at it to determine if it was being asked to
perform FSB delivery. The main problems with this are
 * MSIs are not required anymore to be in any particular vector range
   so it would be completely legitimate for the user to give an `irq_number`
   not between `MSI_MIN` and `MSI_MAX` and be expecting MSI delivery
 * Despite the underlying HPET implementation supporting IOAPIC deliver
   this interface would not attempt to use it
To correct this the new interface provides a separation between the
delivery vector, and whether or not FSB delivery is to be used. With
this extra information the IOAPIC delivery mode of the underyling driver
can also be used.
This is a breaking API change and adds an additional argument to