sel4utils: Add a preallocating slab allocator
This commit adds a preallocating slab allocator which allows users to
specify an amount of objects to preallocate. The allocator will sort the
objects by size and allocate from one untyped of the total size of all
objects, to improve cache locality and reduce cache conflicts.

The allocator does not implement free.