sel4utils: recalculate stack pointer on spawn
Previously the initial_stack_pointer would be used to calculate the
position of the stack after pushing arguments each time a process was
spawned, then the inital_stack_pointer value is mutated to contain the
new value.

This results in the stack being consumed as a process is restarted over
and over again.

This change calculates the initial stack pointer from the original
stack top to avoid consuming the stack.