sel4bench: helpers for looping through events
This commit adds helpers in sel4bench.h to loop through a set of
events when the number of events you want to read are greater than the
number of counters provided by the platform.

These can be used for the generic events or for a custom set
of events.