libsel4vmm: Remove stateful callback that could be called multiple times
The make_guest_boot_info_continued callback performs stateful allocations and mappings.
If the boot_info frame is split across multiple guest pages this callback will be called
multiple times, and therefore attempt to perform this mapping multiple times. There is
no need for such a complicated callback, so this commit changes the logic to construct
the bootinfo in the original make_guest_boot_info function and then just uses a callback
to memcpy the pieces into the guest