SELFOUR-1053: Invoke VMM layer directly in vpsace_[new/map]_pages()
This change brings the behaviour of vspace_[new/map]_pages() in line with its
expected behaviour as the OS page-level combinatorial allocator -- such an
allocator is expected to directly call on the PMM and VMM for frames and pages
respectively, then map the frames into the pages, and do some book-keeping
before handing out the new vmem to userspace.

The previous implementation of vspace_[new/map]_pages had a call to
vspace_new_reservation which in turn had a call to malloc somewhere internally,
and this was causing a circular call path in our libraries.