libsel4platsupport: Construct environment variable for rootserver TCB cptr
Placing the TCB cptr into an environment variables allows for a generic way of talking
about the cptr of the initial thread in a process. As seL4 tells us by agreement where
our TCBs cptr is, we create the environment variable ourself to give to libc.

As the value of this cptr is currently an `enum`, which we cannot reference from
assembly, we also add a C function wrapper for retrieving it