libsel4platsupport: Construct tls elf program headers for rootservers
For TLS support libc expects to be given, in the auxillery variables, where to find
the ELF program headers of the current executable, so that it can determine the TLS
data and bss regions. As seL4 does not pass the elf program headers to us, we attempt
to construct our own.

The construction of this header requires that some symbols defining the tdata and
tbss sections exist. They do not exist by default, but can be provided by a custom
linker script. In the case where they are not provided weak versions of these symbols
are also provided to prevent link errors. When these weak versions get used they
result in the code believing that the tdata and tbss sections are zero sized,
effectively disabling TLS support.