libsel4muslcsys: Call TCB_SetTLSBase when possible
This commit provides support for initial thread in a process, that calls __libc_start_main,
to get a correctly setup thread pointer. If the boot_tcb_cptr environment variable is
set either by the proccess creator, or in the case of the rootserver it will be constructed
by _sel4_start, then during libc's init process, prior to main being called, the
boot_sys_set_thread_area will attempt to set the thread pointer.

Should boot_tcb_cptr not be set an application could still use
muslcsys_get_boot_set_thread_area and then do the equivalent of the code in t his
commit, but this provides a much nicer approach.