SELFOUR-1091: Give hardware more time to print
If we print output too close to when we start the test process, and the
test process resets the serial hardware (which some tests do), our
output characters get deleted from the serial hardware buffers.  (On TK1
and Zync).  The current workaround is to not print anything too close to
when the test process is started.