Make testcase aligned on both 32- and 64-bit
testcases are declared in a specical section and must be aligned to
exactly bytes bits for loops iterating through these sections to work.

With the change to C89 style strings, tests became unaligned on 32bit.
This fixes that problem by defining the string length to the exact
amount to pad the struct to 64 bytes.