Modify behaviour of __arch_* weak symbols
A config option is provided to make it possible to override the
__arch_putchar and __arch_write definitions in libsel4platsupport with
custom ones.  The way this was implemented was by adding a WEAK
attribute to the symbol declaration in the relevant header file that was
guarded by the config variable as well as a WEAK attribute to the symbol
definition that is always set. The limitation of this is that if a
strong symbol definition is not provided at all, the weak symbol won't
be used.

Changing the declaration to always be a strong symbol by removing the
WEAK attribute and moving the config option guards to the
libsel4platsupport definiton of the symbol means that when the
configuration for weak symbols is set, but no strong implementation is
provided then the weak definition can still be used.

One consequence is if the config is not set, two strong definitions will exist and there
will be a linking error whereas before the libsel4platsupport
definition would have been overridden silently.