Add sel4muslcsys_register_stdio_write_fn
This function is for overriding what happens when sys_writev is called
on stdout or stderr.  The default stdio_write function is
sys_platform_write which calls __arch_write defined in libplatsupport
and assumes a rootserver environment.
sel4muslcsys_register_stdio_write_fn should be used in non roottask
environments to specify a custom write function as required.