sel4utils: Add comments and fields to elf_region_t
This structure is for tracking the underlying reservation for an elf
segment.  However because segments can overlap the same frame, we need
to be more careful about how to handle reservation layout.  The new
fields make it possible to keep track of the actual elf segment layout
and also adjustments for the underlying reservation.

When loading the elf data into a vspace it is necessary to have all the
regions because now the start or end of a segment may be in a
reservation belonging to an adjacent region.  The convention is that
all the regions will exist in an ordered array to handle these
conditions.