SELFOUR-1221: Support elf segment frame sharing
It is possible for the linker to generate segments that both overlap the
same frame.  The elf loader now handles this by being aware that the
vspace reservation for its segment may overlap with other adjacent
segments.

It still tracks an individual reservation for each segment. But the
first of last frame of its segment may have been stored in the adjacent
segment's reservation.  The frame gets given to the reservation with the
more permissive access rights.  If the rights are the same, then the
frame is given to the segment with the lower address.  When loading or
manipulating these loadable sections, regions need to be passed around
with their entire array so that the functions can look into adjacent
regions to use their reservations if necessary.