libsel4utils: Remove long running cnode cap deletion
Walking the entire cnode and deleting call the caps was done to ensure there were no
self references left that would keep the object alive. A more efficient way is to
revoke the cnode capability, which will delete any children, allowing the cnode to
be cleaned up by the kernel instead.