New I/O port interface in simple
This changes the simple interface for getting I/O port capabilities to instead of returning
a potentially pre-existing capability for an I/O port, the caller provides a location in
a cnode to place the requested I/O port. The reason for this change is that I/O ports will
soon be being provided from the kernel by a control cap and this design keeps simple as an
information interface that does not need to do internal allocation. Rather allocation is
only performed explicitly by the caller.