Always ensure that some bytes will be written to stack.
The existing implementation of this loop would, whenever the write was
to the start of a new page, write zero bytes (and loop infinitely) to
the page due to rounding down a page boundary to zero.

The new logic ensures that the number of bytes to write will be the
maximum number that can be written before the start of the next page (a
value between 1 and `PAGE_SIZE_4K` inclusive).