libsel4platsupport: Update 'new_io_ops'
This commit updates the 'new_io_ops' function in regards to the addition
of the IRQ interfaces in the I/O ops interface. An additional call has
been added to initialise the IRQ interface with the I/O mapper and
malloc interface. Also add in clean-up functionality when some of the
interfaces fail to initialise.