libsel4platsupport: Fix 'new_fdt_ops'
Previously, we assumed that not being able to copy the FDT was a failure
condition. However, on some platforms, this is actually a valid
condition, e.g. x86 or platforms for which the kernel does not have a
DTS for. This commit fixes that assumption so that calling 'fdt_get'
with the interface returns a NULL pointer.