sel4vspace: Deferred rights reservation mapping
This extends the sel4vspace interface to support the ability to
create reservations where the rights for the memory region are
deferred to the vspace mapping invocation. This is useful when the
creation of a reservation and the mapping of the reservation are
not local and the rights cannot be determined during reservation
creation.