libsel4simple-default: Remove error message
The error message from the 'get_bootinfo_length' function is removed as
this will cause problems for certain projects in which they initialise
the IO ops interface to initialise the serial device.

The problem stems from the fact that the DTB is not included for certain
architectures: x86 and RISCV (for now), and so when the FDT interface
initialisation function calls 'get_bootinfo_length', it will fail to
find the DTB bootinfo section and hence try to print the error message.