sel4muslcsys: Set -u __vsyscall_ptr on link args
Projects that use this library expect __vsyscall_ptr to be present in
the final binary, but require forcing the symbol because it is often not
used by the final binary but by the loader when setting up the AUX
vectors. Adding -u __vsyscall_ptr to the target_link_libraries property
of this library means that the right linker flags will be automatically
provided and applications don't need to keep providing this setting
themselves.