libsel4debug: Correct aarch32 registers names
Somehow along the way, some of the names of the registers of the
kernel's user context for aarch32 got changed and also some other
registers got added. This file wasn't kept updated in regards to user
context change. This commit updates it so that it reflects the user
context changes for aarch32.

Signed-off-by: Damon Lee <Damon.Lee@data61.csiro.au>