libsel4muslcsys,mmap: Assert page aligned value
mmap creates mappings in multiples of the page size which is always at
least 4096. Assert that the return value is at least page size aligned.
Callers of mmap are allowed to assume that the memory returned is 4k
aligned and this can lead to subtle errors if non-aligned values are
returned.

Signed-off-by: Kent McLeod <kent@kry10.com>