muslcsys: print warning when munmap is called
This commit registers a __NR_munmap call, which prevents the
"libsel4muslcsys: Error attempting syscall 215" error from printing.
This syscall is called when a large malloc'd region is freed, since mmap
uses a bump allocator where free is never expected to be called. Instead
of printing a generic warning, a descriptive error is printed which can
alert the user of any potential issues.

Signed-off-by: Chris Guikema <chris.guikema@dornerworks.com>