libsel4test: check err type against unsigned char
On arm-none-elf-eabi-gcc, -fshort-enums is on by default and
is compatible with unsigned char (but not *signed* char) so we
need to add support for this to the test_op macros.

Also, _test_abort() when these type failures happen so that the
failure is more obviously an issue with the compiler/test setup,
not the tests themselves.

Signed-off-by: julia <git.ts@trainwit.ch>